+++ R/content.R
@@ -171,6 +171,7 @@ ContentImageInline <- new_class(
   )
 )
 method(format, ContentImageInline) <- function(x, ...) {
+  show_image(x)
   cli::format_inline("[{.strong inline image}]")
 }
 method(contents_html, ContentImageInline) <- function(content) {
@@ -178,6 +179,17 @@ method(contents_html, ContentImageInline) <- function(content) {
 }
 method(contents_markdown, ContentImageInline) <- function(content) {
   sprintf('![](data:%s;base64,%s)', content@type, content@data)
+}
+
+show_image <- function(x) {
+  if (x@type != "image/png" || !is_installed("png")) {
+    return(invisible())
+  }
+
+  data <- jsonlite::base64_dec(x@data)
+  png <- png::readPNG(data)
+  grid::grid.newpage()
+  grid::grid.raster(png)
 }
 
 # Tools ------------------------------------------------------------------
@@ -324,19 +336,36 @@ tool_string <- function(x, force = FALSE) {
 ContentJson <- new_class(
   "ContentJson",
   parent = Content,
-  properties = list(value = class_any)
+  properties = list(
+    # Some providers guarantee JSON returning it as part of the parsed results.
+    # Others return as a string, which may or may not be valid JSON. We don't
+    # want to force conversion on construction() since that's likely to be
+    # inconveniently early
+    data = class_any,
+    string = prop_string(allow_null = TRUE),
+
+    # Virtual property that always returns parsed JSON or errors
+    parsed = new_property(getter = function(self) {
+      if (is.null(self@string)) {
+        self@data
+      } else {
+        jsonlite::parse_json(self@string)
+      }
+    })
+  )
 )
+
 method(format, ContentJson) <- function(x, ...) {
   paste0(
     cli::format_inline("[{.strong data}] "),
-    pretty_json(x@value)
+    pretty_json(x@parsed)
   )
 }
 method(contents_html, ContentJson) <- function(content) {
-  sprintf('<pre><code>%s</code></pre>\n', pretty_json(content@value))
+  sprintf('<pre><code>%s</code></pre>\n', pretty_json(content@parsed))
 }
 method(contents_markdown, ContentJson) <- function(content) {
-  sprintf('```json\n%s\n```\n', pretty_json(content@value))
+  sprintf('```json\n%s\n```\n', pretty_json(content@parsed))
 }
 
 ContentUploaded <- new_class(