+++ R/provider-ollama.R
@@ -28,6 +28,7 @@
 #'   However, if you're accessing an Ollama instance hosted behind a reverse
 #'   proxy or secured endpoint that enforces bearer‐token authentication, you
 #'   can set `api_key` (or the `OLLAMA_API_KEY` environment variable).
+#' @param params Common model parameters, usually created by [params()].
 #' @inherit chat_openai return
 #' @family chatbots
 #' @export
@@ -40,7 +41,7 @@ chat_ollama <- function(
   system_prompt = NULL,
   base_url = Sys.getenv("OLLAMA_BASE_URL", "http://localhost:11434"),
   model,
-  seed = NULL,
+  params = NULL,
   api_args = list(),
   echo = NULL,
   api_key = NULL,
@@ -73,7 +74,7 @@ chat_ollama <- function(
     name = "Ollama",
     base_url = file.path(base_url, "v1"), ## the v1 portion of the path is added for openAI compatible API
     model = model,
-    seed = seed,
+    params = params %||% params(),
     extra_args = api_args,
     # ollama doesn't require an API key for local usage, but one might be needed
     # if ollama is served behind a proxy (see #501)
@@ -89,10 +90,25 @@ ProviderOllama <- new_class(
   parent = ProviderOpenAI,
   properties = list(
     prop_redacted("api_key"),
-    model = prop_string(),
-    seed = prop_number_whole(allow_null = TRUE)
+    model = prop_string()
   )
 )
+
+method(chat_params, ProviderOllama) <- function(provider, params) {
+  # https://github.com/ollama/ollama/blob/main/docs/api.md#generate-a-chat-completion
+  standardise_params(
+    params,
+    c(
+      frequency_penalty = "frequency_penalty",
+      presence_penalty = "presence_penalty",
+      seed = "seed",
+      stop = "stop_sequences",
+      temperature = "temperature",
+      top_p = "top_p",
+      max_tokens = "max_tokens"
+    )
+  )
+}
 
 chat_ollama_test <- function(..., model = "qwen3:4b", echo = "none") {
   # model: Note that tests require a model with tool capabilities
@@ -130,22 +146,21 @@ models_ollama <- function(base_url = "http://localhost:11434") {
     id = names,
     created_at = modified_at,
     size = size,
-    capabilities = ollama_model_capabilities(names)
+    capabilities = ollama_model_capabilities(base_url, names)
   )
   df[order(-xtfrm(df$created_at)), ]
 }
 
 the$ollama_cache <- new_environment()
 
-ollama_model_details <- function(model) {
+ollama_model_details <- function(base_url, model) {
   # https://github.com/ollama/ollama/blob/main/docs/api.md#show-model-information
-  url <- "http://localhost:11434/api/show"
-
   if (env_has(the$ollama_cache, model)) {
     return(the$ollama_cache[[model]])
   }
 
-  req <- request(url)
+  req <- request(base_url)
+  req <- req_url_path_append(req, "api/show")
   req <- req_body_json(req, list(model = model, verbose = FALSE))
 
   resp <- req_perform(req)
@@ -157,9 +172,9 @@ ollama_model_details <- function(model) {
   details
 }
 
-ollama_model_capabilities <- function(models) {
+ollama_model_capabilities <- function(base_url, models) {
   res <- map(models, function(m) {
-    tryCatch(ollama_model_details(m), error = function(e) NULL)
+    tryCatch(ollama_model_details(base_url, m), error = function(e) NULL)
   })
   map_chr(res, \(x) paste(x$capabilities, collapse = ","))
 }
@@ -177,7 +192,11 @@ has_ollama <- function(base_url = "http://localhost:11434") {
   )
 }
 
-method(as_json, list(ProviderOllama, TypeObject)) <- function(provider, x) {
+method(as_json, list(ProviderOllama, TypeObject)) <- function(
+  provider,
+  x,
+  ...
+) {
   if (x@additional_properties) {
     cli::cli_abort("{.arg .additional_properties} not supported for Ollama.")
   }
@@ -188,7 +207,7 @@ method(as_json, list(ProviderOllama, TypeObject)) <- function(provider, x) {
   compact(list(
     type = "object",
     description = x@description %||% "",
-    properties = as_json(provider, x@properties),
+    properties = as_json(provider, x@properties, ...),
     required = as.list(names2(x@properties)[required]),
     additionalProperties = FALSE
   ))